Add files via upload
Just added a Data engineering pipeline step to the project of fetching the data from the csv file and loading it to the sqlite3. Though in real time the data is fetched from API and some times raw data comes in everyday or depends what we are fetching the data for. In that scenario data is fetched and stored in a database.